Refactor/Suggestion: Change HomeState classes into a single state object.
I think keeping loading, success and failure are useful with the when operator in Kotlin.
The latest documentation on Android suggests using a single state for the UI (similar to MVI and Redux). They are useful in Compose and I think can be useful in Kotlin, as we can define conditionals inside a Stack, Column etc to show a widget, for example, we could use a conditional on isLoading to either show or not this widget.